src/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
distro==1.9.0
exceptiongroup==1.2.2
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.4
importlib_metadata==8.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.8.0
MarkupSafe==2.1.1
openai==1.56.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
requests==2.28.1
sniffio==1.3.1
SQLAlchemy==1.4.42
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.21.0
